08.04.2024 Update

+++ update.sh
@@ -16,20 +16,33 @@
 # https://cvc.keybase.pub/ctlos_repo
 
 local_repo=${PWD}
+arch=x86_64
 echo $local_repo
 repo_ctlos=cretm@cloud.ctlos.ru:/home/cretm/app/cloud.ctlos.ru/ctlos_repo/
 repo_osdn=creio@storage.osdn.net:/storage/groups/c/ct/ctlos/ctlos_repo/
 repo_keybase=/run/user/1000/keybase/kbfs/public/cvc/ctlos_repo/
 
+aur_ctlos=/media/files/github/ctlos/ctlos-aur/x86_64/
+aur_keybase=/run/user/1000/keybase/kbfs/public/cvc/ctlos-aur/
+
+srht_repo=/media/files/srht/ctlos/ctlos_repo/
+
+_git_up() {
+  git add --all
+  msg="$(date +%d.%m.%Y) Update"
+  git commit -a -m "$msg"
+  git push
+}
+
 _keybase() {
   srv_keybase="$(systemctl status --user kbfs | grep -i running 2>/dev/null || echo '')"
   if [[ "$srv_keybase" ]]; then
-    rsync -cauvCLP --delete-excluded --delete --exclude={"build",".git*",".*ignore"} "$local_repo"/ "$repo_keybase"
+    rsync -cauvCLP --delete-excluded --delete --exclude={"build",".git*",".*ignore"} "$local_repo/$arch/" "$repo_keybase"
   else
     systemctl start --user kbfs
     sleep 4
     echo "systemctl start --user kbfs done"
-    rsync -cauvCLP --delete-excluded --delete --exclude={"build",".git*",".*ignore"} "$local_repo"/ "$repo_keybase"
+    rsync -cauvCLP --delete-excluded --delete --exclude={"build",".git*",".*ignore"} "$local_repo/$arch/" "$repo_keybase"
   fi
   # if read -re -p "stop keybase user service? [Y/n]: " ans && [[ $ans == 'n' || $ans == 'N' ]]; then
   #   echo "skip stop kbfs"
@@ -40,7 +53,17 @@ _keybase() {
   echo "rsync keybase repo"
 }
 
+-srht() {
+  if [[ -d "$srht_repo" ]]; then
+    rsync -avrCLP --delete --exclude={"build",".git*"} "$local_repo" "$srht_repo"
+    cd $srht_repo
+    _git_up
+    cd $local_repo
+  fi
+}
+
 if [ "$1" = "-add" ]; then
+  cd $local_repo/$arch
   # repo-add -s -v -n -R ctlos_repo.db.tar.zst *.pkg.tar.xz
   # repo-add -n -R ctlos_repo.db.tar.zst *.pkg.tar.{xz,zst}
   repo-add -n -R -q ctlos_repo.db.tar.zst *.pkg.tar.zst 2>/dev/null;
@@ -54,38 +77,29 @@ elif [ "$1" = "-clean" ]; then
   rm ctlos_repo*
   echo "Repo clean"
 elif [ "$1" = "-o" ]; then
-  rsync -cauvCLP --delete-excluded --delete "$local_repo" "$repo_osdn"
+  # rsync -cauvCLP --delete-excluded --delete "$local_repo/$arch/" "$repo_osdn"
   echo "rsync osdn repo"
-# systemctl --user start kbfs
 elif [ "$1" = "-sync" ]; then
   _keybase
-  rsync -cauvCLP --delete-excluded --delete "$local_repo" "$repo_osdn"
+  _git_up
+  _srht
   echo "rsync all repo"
-# systemctl --user start kbfs
 elif [ "$1" = "-k" ]; then
   _keybase
 elif [ "$1" = "-all" ]; then
+  cd $local_repo/$arch
   repo-add -n -R -q ctlos_repo.db.tar.zst *.pkg.tar.zst
   rm ctlos_repo.{db,files}
   cp -f ctlos_repo.db.tar.zst ctlos_repo.db
   cp -f ctlos_repo.files.tar.zst ctlos_repo.files
   _keybase
-  # rsync -cauvCLP --delete-excluded --delete "$local_repo" "$repo_ctlos"
-  rsync -cauvCLP --delete-excluded --delete "$local_repo" "$repo_osdn"
+  _git_up
+  _srht
   echo "add pkg, rsync all repo"
-else
-  repo-add -n -R ctlos_repo.db.tar.zst *.pkg.tar.zst
-  rm ctlos_repo.{db,files}
-  cp -f ctlos_repo.db.tar.zst ctlos_repo.db
-  cp -f ctlos_repo.files.tar.zst ctlos_repo.files
-  echo "Done repo-add pkg"
 fi
 
 ## sync ctlos-aur repo
-aur_ctlos=cretm@cloud.ctlos.ru:/home/cretm/app/cloud.ctlos.ru/ctlos-aur/
-aur_keybase=/run/user/1000/keybase/kbfs/public/cvc/ctlos-aur/
-
-if [ "$1" = "-aur" ]; then
+if [ "$1" = "-aur" && -d "$aur_ctlos" ]; then
   srv_keybase="$(systemctl status --user kbfs | grep -i running 2>/dev/null || echo '')"
   if [[ "$srv_keybase" ]]; then
     rsync -cauvCLP --delete-excluded --delete "$aur_ctlos" "$aur_keybase"
